+++ src/layout/Nav/Nav.scss
@@ -49,6 +49,18 @@ nav {
     display: flex;
   }
 
+  .notification-count {
+    width: 20px;
+    height: 20px;
+    border-radius: 50%;
+    background-color: rgb(221, 41, 17);
+    position: absolute;
+    top: 23px;
+    display: flex;
+    justify-content: center;
+    font-weight: 700;
+  }
+
   .image {
     background-color: #c4c4c4;
     border-radius: 100%;
